Move Proto->Trace namespace
