+++ lib/opencensus/trace/formatters/binary.rb
@@ -12,12 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "opencensus/proto/trace/annotation"
-require "opencensus/proto/trace/attributes"
-require "opencensus/proto/trace/link"
-require "opencensus/proto/trace/links"
-require "opencensus/proto/trace/message_event"
-require "opencensus/proto/trace/span"
-require "opencensus/proto/trace/status"
-require "opencensus/proto/trace/time_event"
-require "opencensus/proto/trace/time_events"
+module OpenCensus
+  module Trace
+    module Formatters
+      class Binary
+        def deserialize binary
+          
+        end
+
+        def serialize span_context
+        end
+      end
+    end
+  end
+end
